# .restless/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Benefits API",
    "description": "HR and benefits administration API for managing employees, payroll, benefits enrollment, time-off requests, and company documents.",
    "version": "2.1.0"
  },
  "servers": [
    {
      "url": "https://example.com"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/v1/account": {
      "get": {
        "tags": ["Account"],
        "summary": "Get current account",
        "description": "Returns the authenticated account's ID, name, and associated company details.",
        "operationId": "getAccount",
        "responses": {
          "200": {
            "description": "Account details",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "accountId": { "type": "string", "example": "usr_01" },
                    "name": { "type": "string", "example": "Walter Sobchak" },
                    "company": { "$ref": "#/components/schemas/Company" }
                  },
                  "required": ["accountId", "name"]
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
    "/v1/company": {
      "get": {
        "tags": ["Account"],
        "summary": "Get company",
        "description": "Returns the company associated with the authenticated account.",
        "operationId": "getCompany",
        "responses": {
          "200": {
            "description": "Company details",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Company" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      },
      "put": {
        "tags": ["Account"],
        "summary": "Update company",
        "description": "Updates the company associated with the authenticated account. Accepts a partial object with fields to update.",
        "operationId": "updateCompany",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "legalName": { "type": "string" },
                  "dba": { "type": "string" },
                  "ein": { "type": "string" },
                  "addressLine1": { "type": "string" },
                  "city": { "type": "string" },
                  "state": { "type": "string" },
                  "postalCode": { "type": "string" },
                  "plan": { "type": "string" },
                  "payrollFrequency": { "type": "string" },
                  "nextPayDate": { "type": "string", "format": "date" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated company",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Company" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
    "/v1/employees": {
      "get": {
        "tags": ["Employees"],
        "summary": "List employees",
        "description": "Returns all employees belonging to the authenticated account's company. Supports filtering by status, department, and employment type.",
        "operationId": "listEmployees",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "description": "Filter by employee status (e.g. active, terminated)",
            "schema": { "type": "string" }
          },
          {
            "name": "department",
            "in": "query",
            "description": "Filter by department name",
            "schema": { "type": "string" }
          },
          {
            "name": "employmentType",
            "in": "query",
            "description": "Filter by employment type (salary or hourly)",
            "schema": { "type": "string", "enum": ["salary", "hourly"] }
          }
        ],
        "responses": {
          "200": {
            "description": "List of employees",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/Employee" }
                    },
                    "count": { "type": "integer" }
                  },
                  "required": ["data", "count"]
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      },
      "post": {
        "tags": ["Employees"],
        "summary": "Create employee",
        "description": "Creates a new employee in the authenticated account's company. If employmentType is \"salary\", annualSalaryCents is required. If employmentType is \"hourly\", hourlyRateCents is required.",
        "operationId": "createEmployee",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "firstName": { "type": "string" },
                  "lastName": { "type": "string" },
                  "email": { "type": "string", "format": "email" },
                  "startDate": { "type": "string", "format": "date" },
                  "employmentType": { "type": "string", "enum": ["salary", "hourly"] },
                  "annualSalaryCents": { "type": "integer", "description": "Required when employmentType is salary" },
                  "hourlyRateCents": { "type": "integer", "description": "Required when employmentType is hourly" },
                  "title": { "type": "string" },
                  "department": { "type": "string" },
                  "ssnLast4": { "type": "string" },
                  "dob": { "type": "string", "format": "date" }
                },
                "required": ["firstName", "lastName", "email", "startDate", "employmentType"]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Employee created",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Employee" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "422": { "$ref": "#/components/responses/ValidationError" }
        }
      }
    },
    "/v1/employees/{id}": {
      "get": {
        "tags": ["Employees"],
        "summary": "Get employee",
        "description": "Returns a single employee by ID, scoped to the authenticated account's company.",
        "operationId": "getEmployee",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Employee ID",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Employee details",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Employee" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      },
      "put": {
        "tags": ["Employees"],
        "summary": "Update employee",
        "description": "Updates an existing employee by ID. Accepts a partial object with fields to update.",
        "operationId": "updateEmployee",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Employee ID",
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "firstName": { "type": "string" },
                  "lastName": { "type": "string" },
                  "email": { "type": "string", "format": "email" },
                  "title": { "type": "string" },
                  "department": { "type": "string" },
                  "employmentType": { "type": "string", "enum": ["salary", "hourly"] },
                  "annualSalaryCents": { "type": "integer" },
                  "hourlyRateCents": { "type": "integer" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated employee",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Employee" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
    "/v1/employees/{id}/terminate": {
      "post": {
        "tags": ["Employees"],
        "summary": "Terminate employee",
        "description": "Terminates an active employee. Optionally accepts a termination date and reason. Returns a 409 if the employee is already terminated.",
        "operationId": "terminateEmployee",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Employee ID",
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "terminationDate": { "type": "string", "format": "date", "description": "Defaults to today if not provided" },
                  "reason": { "type": "string", "description": "Defaults to \"voluntary\" if not provided" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Employee terminated",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Employee" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "409": {
            "description": "Employee is already terminated",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      }
    },
    "/v1/employees/{id}/rehire": {
      "post": {
        "tags": ["Employees"],
        "summary": "Rehire employee",
        "description": "Rehires a previously terminated employee, setting their status back to active and clearing termination details.",
        "operationId": "rehireEmployee",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Employee ID",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Employee rehired",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Employee" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
    "/v1/payrolls": {
      "get": {
        "tags": ["Payroll"],
        "summary": "List payrolls",
        "description": "Returns all payroll runs for the authenticated account's company.",
        "operationId": "listPayrolls",
        "responses": {
          "200": {
            "description": "List of payrolls",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/Payroll" }
                    },
                    "count": { "type": "integer" }
                  },
                  "required": ["data", "count"]
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      },
      "post": {
        "tags": ["Payroll"],
        "summary": "Create payroll",
        "description": "Creates a new payroll run in draft status. Active employees are automatically included, and amounts are initialized to zero.",
        "operationId": "createPayroll",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "periodStart": { "type": "string", "format": "date" },
                  "periodEnd": { "type": "string", "format": "date" },
                  "payDate": { "type": "string", "format": "date" }
                },
                "required": ["periodStart", "periodEnd", "payDate"]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Payroll created",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Payroll" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "422": { "$ref": "#/components/responses/ValidationError" }
        }
      }
    },
    "/v1/payrolls/{id}": {
      "get": {
        "tags": ["Payroll"],
        "summary": "Get payroll",
        "description": "Returns a single payroll run by ID, scoped to the authenticated account's company.",
        "operationId": "getPayroll",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Payroll ID",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Payroll details",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Payroll" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
    "/v1/payrolls/{id}/submit": {
      "post": {
        "tags": ["Payroll"],
        "summary": "Submit payroll",
        "description": "Submits a draft payroll for processing. Returns a 409 if the payroll has already been paid.",
        "operationId": "submitPayroll",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Payroll ID",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Payroll submitted for processing",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Payroll" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "409": {
            "description": "Payroll has already been paid",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      }
    },
    "/v1/payrolls/{id}/cancel": {
      "post": {
        "tags": ["Payroll"],
        "summary": "Cancel payroll",
        "description": "Cancels a payroll run. Returns a 409 if the payroll has already been paid.",
        "operationId": "cancelPayroll",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Payroll ID",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Payroll cancelled",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Payroll" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "409": {
            "description": "Cannot cancel a paid payroll",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      }
    },
    "/v1/benefits": {
      "get": {
        "tags": ["Benefits"],
        "summary": "List benefits",
        "description": "Returns all available benefit plans (medical, dental, vision, retirement, etc.).",
        "operationId": "listBenefits",
        "responses": {
          "200": {
            "description": "List of benefits",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/Benefit" }
                    }
                  },
                  "required": ["data"]
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Pass your API token via the Authorization header: `Authorization: Token <token>` or `Authorization: Bearer <token>`"
      }
    },
    "schemas": {
...
      "Company": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "example": "cmp_01" },
          "legalName": { "type": "string", "example": "Sobchak Security LLC" },
          "dba": { "type": "string", "example": "Sobchak Security" },
          "ein": { "type": "string", "example": "85-2244719" },
          "addressLine1": { "type": "string", "example": "608 Venezia Ave" },
          "city": { "type": "string", "example": "Venice" },
          "state": { "type": "string", "example": "CA" },
          "postalCode": { "type": "string", "example": "90291" },
          "plan": { "type": "string", "example": "complete" },
          "payrollFrequency": { "type": "string", "example": "biweekly" },
          "nextPayDate": { "type": "string", "format": "date", "example": "2026-05-22" }
        },
        "required": ["id", "legalName"]
      },
...
    },
    "responses": {
      "Unauthorized": {
        "description": "Missing or invalid API token",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" }
          }
        }
      },
...
    }
  }
}